invoker/requirements.txt
protobuf==3.4.0
futures==3.2.0
grpcio==1.8.4